requirements.txt
Flask==3.1.1
python-dotenv==1.0.1
google-generativeai==0.3.2
blinker==1.9.0
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3